dev-1.3.6.7
Soulmates in Double Life no longer get damage ticked when you take damage while blocking with a shield.
Improved the '/boogeyman list' command.

+++ src/main/java/net/mat0u5/lifeseries/seasons/boogeyman/BoogeymanManager.java
@@ -140,6 +140,7 @@ public class BoogeymanManager {
         if (!BOOGEYMAN_ENABLED) return;
         Boogeyman boogeyman = getBoogeyman(player);
         if (boogeymen == null) return;
+        boogeyman.failed = false;
         boogeyman.cured = true;
         PlayerUtils.sendTitle(player,Text.of("§aYou are cured!"), 20, 30, 20);
         PlayerUtils.playSoundToPlayers(List.of(player), SoundEvent.of(Identifier.of("minecraft","lastlife_boogeyman_cure")));
@@ -258,12 +259,16 @@ public class BoogeymanManager {
 
     public void playerFailBoogeyman(ServerPlayerEntity player) {
         if (!BOOGEYMAN_ENABLED) return;
+        Boogeyman boogeyman = getBoogeyman(player);
+        if (boogeymen == null) return;
         if (!currentSeason.isAlive(player)) return;
         if (currentSeason.isOnLastLife(player, true)) return;
         PlayerUtils.sendTitle(player,Text.of("§cYou have failed."), 20, 30, 20);
         PlayerUtils.playSoundToPlayers(List.of(player), SoundEvent.of(Identifier.of("minecraft","lastlife_boogeyman_fail")));
         OtherUtils.broadcastMessage(player.getStyledDisplayName().copy().append(Text.of("§7 failed to kill a player while being the §cBoogeyman§7. They have been dropped to their §cLast Life§7")));
         currentSeason.setPlayerLives(player, 1);
+        boogeyman.failed = true;
+        boogeyman.cured = false;
         if (currentSeason instanceof DoubleLife doubleLife) {
             doubleLife.syncSoulboundLives(player);
         }
